# MalyFarmar/MalyFarmar.Api/ClientGeneration/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MalyFarmar API",
    "version": "v1"
  },
  "paths": {
    "/api/Order/{orderId}": {
      "get": {
        "tags": [
          "Order"
        ],
        "operationId": "GetOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailViewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailViewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailViewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/create": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "CreateOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailViewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailViewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailViewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/{orderId}/set-pick-up-date-time": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "SetPickUpDateTime",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSetPickUpDateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSetPickUpDateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OrderSetPickUpDateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Order/{orderId}/set-order-completed": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "SetOrderCompleted",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Order/get-orders/{buyerId}": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "GetOrders",
        "parameters": [
          {
            "name": "buyerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Order/ger-reservations/{sellerId}": {
      "post": {
        "tags": [
          "Order"
        ],
        "operationId": "GetReservations",
        "parameters": [
          {
            "name": "sellerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/{productId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "GetProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "operationId": "DeleteProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Product/get-products": {
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "GetProducts",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductSearchDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductSearchDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductSearchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/get-products-by-seller/{sellerId}": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "GetProductsBySeller",
        "parameters": [
          {
            "name": "sellerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductsListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/create": {
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "CreateProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Product/{productId}/update": {
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "UpdateProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductEditDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductEditDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductEditDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailViewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserViewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserViewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserViewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/get-all": {
      "get": {
        "tags": [
          "User"
        ],
        "operationId": "GetAllUsers",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/create": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "CreateUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserViewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserViewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserViewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/{userId}/set-location": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "SetUserLocation",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSetLocationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSetLocationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserSetLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "OrderCreateDto": {
        "required": [
          "buyerId",
          "items"
        ],
        "type": "object",
        "properties": {
          "buyerId": {
            "type": "integer",
            "format": "int32"
          },
          "pickUpAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemCreateDto"
            }
          }
        },
        "additionalProperties": false
      },
      "OrderDetailViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "pickUpAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "statusId": {
            "$ref": "#/components/schemas/OrderStatusEnum"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDetailViewDto"
            },
            "nullable": true
          },
          "seller": {
            "$ref": "#/components/schemas/UserViewDto"
          },
          "buyer": {
            "$ref": "#/components/schemas/UserViewDto"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "OrderItemCreateDto": {
        "required": [
          "amount",
          "pricePerUnit",
          "productId"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OrderItemDetailViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          },
          "product": {
            "$ref": "#/components/schemas/ProductDetailViewDto"
          }
        },
        "additionalProperties": false
      },
      "OrderListViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "pickUpAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "statusId": {
            "$ref": "#/components/schemas/OrderStatusEnum"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "OrderSetPickUpDateDto": {
        "required": [
          "pickUpAt"
        ],
        "type": "object",
        "properties": {
          "pickUpAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "OrderStatusEnum": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "OrdersListDto": {
        "type": "object",
        "properties": {
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderListViewDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductCreateDto": {
        "required": [
          "name",
          "pricePerUnit",
          "sellerId",
          "totalAmount",
          "unit"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "minLength": 1,
            "type": "string"
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          },
          "sellerId": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductDetailViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "remainingAmount": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          },
          "seller": {
            "$ref": "#/components/schemas/UserViewDto"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductEditDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProductListViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "remainingAmount": {
            "type": "number",
            "format": "double"
          },
          "unit": {
            "type": "string",
            "nullable": true
          },
          "pricePerUnit": {
            "type": "number",
            "format": "double"
          },
          "distanceInMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductSearchDto": {
        "required": [
          "latitude",
          "longitude"
        ],
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "radiusInMeters": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductsListDto": {
        "type": "object",
        "properties": {
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductListViewDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserCreateDto": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string"
          },
          "phoneNumber": {
            "minLength": 1,
            "type": "string"
          },
          "userLongitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "userLatitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserListViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserSetLocationDto": {
        "required": [
          "userLatitude",
          "userLongitude"
        ],
        "type": "object",
        "properties": {
          "userLongitude": {
            "type": "number",
            "format": "double"
          },
          "userLatitude": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "UserViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "userLongitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "userLatitude": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UsersListDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserListViewDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}